+++ src/main/java/enums/Direction.java
@@ -1,4 +1,4 @@
-package Enums;
+package enums;
 
 public enum Direction {
     UP, DOWN, LEFT, RIGHT;
