+++ src/main/java/inf112/skeleton/app/Board.java
@@ -6,7 +6,7 @@ import com.badlogic.gdx.graphics.g2d.*;
 import com.badlogic.gdx.maps.tiled.*;
 import com.badlogic.gdx.maps.tiled.renderers.*;
 import com.badlogic.gdx.maps.tiled.tiles.*;
-import Class.*;
+import classes.*;
 
 public class Board extends InputAdapter implements ApplicationListener {
