+++ src/main/java/interfaces/ICard.java
@@ -1,4 +1,4 @@
-package Interface;
+package interfaces;
 
 public interface ICard {
