+++ src/main/java/interfaces/ICell.java
@@ -1,6 +1,6 @@
-package Interface;
+package interfaces;
 
-import Enums.Direction;
+import enums.Direction;
 
 public interface ICell {
     boolean hasWall(Direction dir);
