+++ src/main/java/interfaces/IHand.java
@@ -1,4 +1,4 @@
-package Interface;
+package interfaces;
 
 import java.util.List;
