+++ src/main/java/interfaces/IMapPiece.java
@@ -1,4 +1,4 @@
-package Interface;
+package interfaces;
 
 public interface IMapPiece {
