+++ src/main/java/interfaces/IPosition.java
@@ -1,4 +1,4 @@
-package Interface;
+package interfaces;
 
 public interface IPosition {
